renamed and refactoed SquareTile to TileImage. Leaner and meaner but basically the same.
